+++ Assets/Scripts/Application Layer/TownSystem/LogProcessingSystem/LogContainer/LogContainer.cs
@@ -27,9 +27,9 @@ public class LogContainer : MonoBehaviour, IInventory, IContainerCH
 
     IReadOnlyList<IInventorySlot> IInventory.inventorySlots => containerSlots;
 
-    public int money => 0;
+    public long money => 0;
 
-    public int carrot => 0;
+    public long carrot => 0;
 
     public int currentSlotCnt => currentSlotCount;
