+++ Assets/Scripts/Application Layer/UnitSystem/Interface/IInventory.cs
@@ -7,6 +7,6 @@ public interface IInventory
     IReadOnlyList<IInventorySlot> inventorySlots { get; }
     int currentSlotCnt { get; }
     Transform GetTransform();
-    int money { get; }
-    int carrot { get; }
+    long money { get; }
+    long carrot { get; }
 }
